requirements.txt
python-dotenv 
requests 
aiohttp
groq
twilio
uvicorn
pyngrok
fastapi
ipykernel
pyTelegramBotAPI
mistralai
google-genai
Pillow